chore: fix studio-connectors imports
