+++ packages/connectors/package.json
@@ -11,8 +11,8 @@
         "prettier": "^3.0.3",
         "typescript": "^5.2.2",
         "vite": "^6.3.5",
-        "vite-plugin-dts": "^4.5.3",
-        "@types/node": "^22.15.18"
+        "@types/node": "^22.15.18",
+        "dts-bundle-generator": "^9.0.0"
     },
     "scripts": {
         "build": "vite build",
