+++ packages/connectors/src/MediaConnector.ts
@@ -27,6 +27,6 @@ export type DownloadIntent = 'web' | 'print' | 'animation';
 export type DownloadType = 'thumbnail' | 'mediumres' | 'highres' | 'fullres' | 'original';
 export type MediaType = 0 | 1;
 
-type MediaPage = QueryPage<Media>;
+export type MediaPage = QueryPage<Media>;
 export type Media = GenericMedia<MediaType>;
 export type MediaDetail = GenericMediaDetail<MediaType>;
